+++ src/sdhlt/common/bspfile.h
@@ -363,6 +363,7 @@ extern bool		CalcFaceExtents_test ();
 extern void		LoadExtentFile (const char *const filename);
 #endif
 extern void		GetFaceExtents (int facenum, int mins_out[2], int maxs_out[2]);
+extern int		ParseTexinfoFromBakedName (const char *name);
 extern int		ParseImplicitTexinfoFromTexture (int miptex);
 extern int		ParseTexinfoForFace (const dface_t *f);
 extern void		DeleteEmbeddedLightmaps ();